requirements.txt
crewai 
crewai_tools 
openai
discord.py
beautifulsoup4
python-dotenv
numpy
langchain
pandas
python-dotenv
pydantic
fake-useragent
ollama
discord.py
scikit-learn
selenium==4.10.0
webdriver-manager==3.8.6

